+++ tests/ryan_joiner/test_dist_plot.py
@@ -3,9 +3,6 @@
 --------------------------------------------------------------------------------
 Command to run at the prompt:
     python -m unittest -v tests/ryan_joiner/test_dist_plot.py
-    or
-    python -m unittest -b tests/ryan_joiner/test_dist_plot.py
-
 --------------------------------------------------------------------------------
 """
 ### GENERAL IMPORTS ###
@@ -27,45 +24,61 @@ class Test_dist_plot(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         fig, cls.axes = plt.subplots()
-        cls.x_data = np.array(
-            [148, 148, 154, 158, 158, 160, 161, 162, 166, 170, 182, 195, 210]
-        )
+        cls.statistic = 0.9225156050800545
+        cls.sample_size = 13
 
     def test_outputs(self):
         result = dist_plot(
             self.axes,
-            self.x_data,
-            cte_alpha="3/8",
-            min=4,
-            max=50,
-            weighted=False,
-            safe=False,
         )
         self.assertIsInstance(result, SubplotBase, msg="not a SubplotBase")
         plt.close()
 
         result = dist_plot(
             axes=self.axes,
-            x_data=self.x_data,
-            cte_alpha="3/8",
-            min=4,
-            max=50,
-            weighted=False,
-            safe=False,
+            critical_range=(4, 50),
+            test=(self.statistic, self.sample_size),
         )
         self.assertIsInstance(result, SubplotBase, msg="not a SubplotBase")
         plt.close()
 
-    def test_safe(self):
-        result = dist_plot(self.axes, self.x_data, safe=True)
-        self.assertIsInstance(result, SubplotBase, msg="not a SubplotBase")
-        plt.close()
-
     def test_basic_plot(self):
-        fig1_base_path = Path("tests/ryan_joiner/figs_dist_plot/fig1.png")
+        fig1_base_path = Path("tests/ryan_joiner/figs_dist_plot/dist_plot_default.png")
 
         fig, ax = plt.subplots()
-        result = dist_plot(ax, self.x_data)
+        result = dist_plot(ax)
+        fig1_file = Path("tests/ryan_joiner/figs_dist_plot/fig1_test.png")
+        plt.savefig(fig1_file)
+        plt.close()
+
+        self.assertTrue(
+            functions.validate_file_contents(fig1_base_path, fig1_file),
+            msg="figures does not match",
+        )
+        fig1_file.unlink()
+
+    def test_plot_with_test(self):
+        fig1_base_path = Path("tests/ryan_joiner/figs_dist_plot/dist_plot_test.png")
+
+        fig, ax = plt.subplots()
+        result = dist_plot(ax, test=(self.statistic, self.sample_size))
+        fig1_file = Path("tests/ryan_joiner/figs_dist_plot/fig1_test.png")
+        plt.savefig(fig1_file)
+        plt.close()
+
+        self.assertTrue(
+            functions.validate_file_contents(fig1_base_path, fig1_file),
+            msg="figures does not match",
+        )
+        fig1_file.unlink()
+
+    def test_plot_critical_range(self):
+        fig1_base_path = Path("tests/ryan_joiner/figs_dist_plot/dist_plot_range.png")
+
+        fig, ax = plt.subplots()
+        result = dist_plot(
+            ax, test=(self.statistic, self.sample_size), critical_range=(5, 30)
+        )
         fig1_file = Path("tests/ryan_joiner/figs_dist_plot/fig1_test.png")
         plt.savefig(fig1_file)
         plt.close()